Attempt #1: Stepper control
Stepper Control is not hard, the hard part is command flow.

For example: you have 2 scripts, one to move a stepper motor clockwise and the other to move it counterclockwise.

How to make it so that each will interrupt and stop the other script when run?

My solution (supposedly) is to change & check the FSEL state of a pin.