Stepper Motor Support Added
Stepper Control implemented

The biggest challenge was to how to change direction a stepper motor?
How to interfere a running python script with a new one?

There needs to be a listener & a state. In this case a physical & electrical state: a pinout state/function.

When changing direction, one pin (pin1) will change from GPIO.OUTPUT to GPIO.INPUT for 0.1 seconds. This is just enough for an if function in the existing running script to catch an error and exit(1).

This results in a perfectly working solution with a 0.1s delay. More improvement on the delay will be made if requested.

Moreover, thanks to Tuyen Nguyen's recommendation, stepper motor script is written with OOP in mind.